+++ js/consts.js
@@ -1,0 +1,42 @@
+export const NAMES = [
+  'Вася пупкин',
+  'Иван Иванов',
+  'Волкова Анастасия'
+];
+
+export const DESCRIPTIONS = [
+  'Отличный день на пляже!',
+  'Горный поход был прекрасен',
+  'Мой новый питомец',
+  'Вкусный ужин в ресторане',
+  'Закат над океаном',
+  'Уютный вечер с книгой',
+  'Путешествие в новый город',
+  'Встреча с друзьями',
+  'Мои цветы в саду',
+  'Зимняя сказка',
+  'Архитектура старого города',
+  'Кофе и работа',
+  'Спортивные достижения',
+  'Романтический вечер',
+  'Фестиваль музыки',
+  'Прогулка в лесу',
+  'Моё хобби',
+  'Семейный праздник',
+  'Новое хобби - фотография',
+  'Уикенд на природе',
+  'Городские огни ночью',
+  'Вкус домашней выпечки',
+  'Утро начинается с кофе',
+  'Осенние краски',
+  'Моменты счастья'
+];
+
+export const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!'
+];
